refactor: rename file storage service
